+++ bindings/angular2/src/ngx-onsenui.ts
@@ -29,6 +29,7 @@ import {OnsSelect} from './directives/ons-select';
 import {OnsInput} from './directives/ons-input';
 import {OnsSearchInput} from './directives/ons-search-input';
 import {OnsCheckbox} from './directives/ons-checkbox';
+import {OnsRadio} from './directives/ons-radio';
 import {OnsPullHook} from './directives/ons-pull-hook';
 import {OnsLazyRepeat} from './directives/ons-lazy-repeat';
 import {OnsSplitterSide, OnsSplitterContent} from './directives/ons-splitter';
@@ -48,6 +49,7 @@ const directives = [
   OnsInput,
   OnsSearchInput,
   OnsCheckbox,
+  OnsRadio,
   OnsPullHook,
   OnsLazyRepeat,
   OnsSplitterSide,